+++ components/logins/src/schema.rs
@@ -295,7 +295,7 @@ pub(crate) fn create(db: &Connection) -> Result<()> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::encryption::test_utils::TEST_ENCDEC;
+    use crate::test_utils::TEST_ENCDEC;
     use crate::LoginDb;
     use nss_as::ensure_initialized;
     use rusqlite::Connection;