+++ components/support/db-crypto/src/encryption.rs
@@ -3,14 +3,14 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-// This is the *local* encryption support - it has nothing to do with the
-// encryption used by sync.
+// This is the *local* database encryption support - it has nothing to do
+// with the encryption used by sync.
 
 // For context, what "local encryption" means in this context is:
 // * We use regular sqlite, but ensure that sensitive data is encrypted in the DB in the
 //   `secure_fields` column.  The encryption key is managed by the app.
-// * The `decrypt_struct` and `encrypt_struct` functions are used to convert between an encrypted
-//   `secure_fields` string and a decrypted `SecureFields` struct
+// * The `decrypt` and `encrypt` functions are used to convert between an encrypted
+//   string and a decrypted string.
 // * Most API functions return `EncryptedLogin` which has its data encrypted.
 //
 // This makes life tricky for Sync - sync has its own encryption and its own
@@ -27,7 +27,7 @@
 // low level sync code.
 // To make life a little easier, we do that via a struct.
 //
-// Consumers of the Login component have 3 options for setting up encryption:
+// Consumers of the an encrypted database have 3 options for setting up encryption:
 //    1. Implement EncryptorDecryptor directly
 //       eg `LoginStore::new(MyEncryptorDecryptor)`
 //    2. Implement KeyManager and use ManagedEncryptorDecryptor
@@ -112,20 +112,20 @@ impl EncryptorDecryptor for ManagedEncryptorDecryptor {
         let keybytes = self
             .key_manager
             .get_key()
-            .map_err(|_| LoginsApiError::MissingKey)?;
-        let key = std::str::from_utf8(&keybytes).map_err(|_| LoginsApiError::InvalidKey)?;
+            .map_err(|_| DbCryptoApiError::MissingKey)?;
+        let key = std::str::from_utf8(&keybytes).map_err(|_| DbCryptoApiError::InvalidKey)?;
 
         let encdec = jwcrypto::EncryptorDecryptor::new(key)
-            .map_err(|_: jwcrypto::JwCryptoError| LoginsApiError::InvalidKey)?;
+            .map_err(|_: jwcrypto::JwCryptoError| DbCryptoApiError::InvalidKey)?;
 
         let cleartext =
-            std::str::from_utf8(&clearbytes).map_err(|e| LoginsApiError::EncryptionFailed {
+            std::str::from_utf8(&clearbytes).map_err(|e| DbCryptoApiError::EncryptionFailed {
                 reason: e.to_string(),
             })?;
         encdec
             .encrypt(cleartext)
             .map_err(
-                |e: jwcrypto::JwCryptoError| LoginsApiError::EncryptionFailed {
+                |e: jwcrypto::JwCryptoError| DbCryptoApiError::EncryptionFailed {
                     reason: e.to_string(),
                 },
             )
@@ -136,20 +136,20 @@ impl EncryptorDecryptor for ManagedEncryptorDecryptor {
         let keybytes = self
             .key_manager
             .get_key()
-            .map_err(|_| LoginsApiError::MissingKey)?;
-        let key = std::str::from_utf8(&keybytes).map_err(|_| LoginsApiError::InvalidKey)?;
+            .map_err(|_| DbCryptoApiError::MissingKey)?;
+        let key = std::str::from_utf8(&keybytes).map_err(|_| DbCryptoApiError::InvalidKey)?;
 
         let encdec = jwcrypto::EncryptorDecryptor::new(key)
-            .map_err(|_: jwcrypto::JwCryptoError| LoginsApiError::InvalidKey)?;
+            .map_err(|_: jwcrypto::JwCryptoError| DbCryptoApiError::InvalidKey)?;
 
         let ciphertext =
-            std::str::from_utf8(&cipherbytes).map_err(|e| LoginsApiError::DecryptionFailed {
+            std::str::from_utf8(&cipherbytes).map_err(|e| DbCryptoApiError::DecryptionFailed {
                 reason: e.to_string(),
             })?;
         encdec
             .decrypt(ciphertext)
             .map_err(
-                |e: jwcrypto::JwCryptoError| LoginsApiError::DecryptionFailed {
+                |e: jwcrypto::JwCryptoError| DbCryptoApiError::DecryptionFailed {
                     reason: e.to_string(),
                 },
             )
@@ -212,36 +212,37 @@ pub trait PrimaryPasswordAuthenticator: Send + Sync {
 /// # Examples
 /// ```no_run
 /// use async_trait::async_trait;
-/// use logins::encryption::KeyManager;
-/// use logins::{PrimaryPasswordAuthenticator, LoginsApiError, NSSKeyManager};
+/// use db_crypto::KeyManager;
+/// use db_crypto::{PrimaryPasswordAuthenticator, DbCryptoApiError, NSSKeyManager};
 /// use std::sync::Arc;
 ///
 /// struct MyPrimaryPasswordAuthenticator {}
 ///
 /// #[async_trait]
 /// impl PrimaryPasswordAuthenticator for MyPrimaryPasswordAuthenticator {
-///     async fn get_primary_password(&self) -> Result<String, LoginsApiError> {
+///     async fn get_primary_password(&self) -> Result<String, DbCryptoApiError> {
 ///         // Most likely, you would want to prompt for a password.
 ///         // let password = prompt_string("primary password").unwrap_or_default();
 ///         Ok("secret".to_string())
 ///     }
 ///
-///     async fn on_authentication_success(&self) -> Result<(), LoginsApiError> {
+///     async fn on_authentication_success(&self) -> Result<(), DbCryptoApiError> {
 ///         println!("success");
 ///         Ok(())
 ///     }
 ///
-///     async fn on_authentication_failure(&self) -> Result<(), LoginsApiError> {
+///     async fn on_authentication_failure(&self) -> Result<(), DbCryptoApiError> {
 ///         println!("this did not work, please try again:");
 ///         Ok(())
 ///     }
 /// }
-/// let key_manager = NSSKeyManager::new(Arc::new(MyPrimaryPasswordAuthenticator {}));
+/// let key_manager = NSSKeyManager::new(String::from("example"), Arc::new(MyPrimaryPasswordAuthenticator {}));
 /// assert_eq!(key_manager.get_key().unwrap().len(), 63);
 /// ```
 #[cfg(feature = "keydb")]
 #[derive(uniffi::Object)]
 pub struct NSSKeyManager {
+    key_name: String,
     primary_password_authenticator: Arc<dyn PrimaryPasswordAuthenticator>,
     cached_key: RwLock<Option<Vec<u8>>>,
 }
@@ -253,9 +254,13 @@ impl NSSKeyManager {
     /// There must be a previous initializiation of NSS before initializing
     /// `NSSKeyManager`, otherwise this panics.
     #[uniffi::constructor()]
-    pub fn new(primary_password_authenticator: Arc<dyn PrimaryPasswordAuthenticator>) -> Self {
+    pub fn new(
+        key_name: String,
+        primary_password_authenticator: Arc<dyn PrimaryPasswordAuthenticator>,
+    ) -> Self {
         assert_nss_initialized();
         Self {
+            key_name,
             primary_password_authenticator,
             cached_key: RwLock::new(None),
         }
@@ -266,15 +271,11 @@ impl NSSKeyManager {
     }
 }
 
-/// Identifier for the logins key, under which the key is stored in NSS.
-#[cfg(feature = "keydb")]
-static KEY_NAME: &str = "as-logins-key";
-
 // wrapp `authentication_with_primary_password_is_needed` into an ApiResult
 #[cfg(feature = "keydb")]
 fn api_authentication_with_primary_password_is_needed() -> ApiResult<bool> {
     authentication_with_primary_password_is_needed().map_err(|e: nss_as::Error| {
-        LoginsApiError::NSSAuthenticationError {
+        DbCryptoApiError::NSSAuthenticationError {
             reason: e.to_string(),
         }
     })
@@ -284,7 +285,7 @@ fn api_authentication_with_primary_password_is_needed() -> ApiResult<bool> {
 #[cfg(feature = "keydb")]
 fn api_authenticate_with_primary_password(primary_password: &str) -> ApiResult<bool> {
     authenticate_with_primary_password(primary_password).map_err(|e: nss_as::Error| {
-        LoginsApiError::NSSAuthenticationError {
+        DbCryptoApiError::NSSAuthenticationError {
             reason: e.to_string(),
         }
     })
@@ -329,7 +330,8 @@ impl KeyManager for NSSKeyManager {
             return Ok(bytes);
         }
 
-        let key = get_or_create_aes256_key(KEY_NAME).map_err(|_| LoginsApiError::MissingKey)?;
+        let key = get_or_create_aes256_key(self.key_name.as_str())
+            .map_err(|_| DbCryptoApiError::MissingKey)?;
         let mut bytes: Vec<u8> = Vec::new();
         serde_json::to_writer(
             &mut bytes,
@@ -348,34 +350,13 @@ pub fn create_canary(text: &str, key: &str) -> ApiResult<String> {
 
 pub fn check_canary(canary: &str, text: &str, key: &str) -> ApiResult<bool> {
     let encdec = jwcrypto::EncryptorDecryptor::new(key)
-        .map_err(|_: jwcrypto::JwCryptoError| LoginsApiError::InvalidKey)?;
+        .map_err(|_: jwcrypto::JwCryptoError| DbCryptoApiError::InvalidKey)?;
     Ok(encdec.check_canary(canary, text).unwrap_or(false))
 }
 
 #[handle_error(Error)]
 pub fn create_key() -> ApiResult<String> {
     Ok(jwcrypto::EncryptorDecryptor::create_key()?)
-}
-
-#[cfg(test)]
-pub mod test_utils {
-    use super::*;
-    use serde::{de::DeserializeOwned, Serialize};
-
-    lazy_static::lazy_static! {
-        pub static ref TEST_ENCRYPTION_KEY: String = serde_json::to_string(&jwcrypto::Jwk::new_direct_key(Some("test-key".to_string())).unwrap()).unwrap();
-        pub static ref TEST_ENCDEC: Arc<ManagedEncryptorDecryptor> = Arc::new(ManagedEncryptorDecryptor::new(Arc::new(StaticKeyManager { key: TEST_ENCRYPTION_KEY.clone() })));
-    }
-
-    pub fn encrypt_struct<T: Serialize>(fields: &T) -> String {
-        let string = serde_json::to_string(fields).unwrap();
-        let cipherbytes = TEST_ENCDEC.encrypt(string.as_bytes().into()).unwrap();
-        std::str::from_utf8(&cipherbytes).unwrap().to_owned()
-    }
-    pub fn decrypt_struct<T: DeserializeOwned>(ciphertext: String) -> T {
-        let jsonbytes = TEST_ENCDEC.decrypt(ciphertext.as_bytes().into()).unwrap();
-        serde_json::from_str(std::str::from_utf8(&jsonbytes).unwrap()).unwrap()
-    }
 }
 
 #[cfg(not(feature = "keydb"))]
@@ -401,7 +382,7 @@ mod tests {
         let encdec = ManagedEncryptorDecryptor { key_manager };
         assert!(matches!(
             encdec.encrypt("secret".as_bytes().into()).err().unwrap(),
-            LoginsApiError::InvalidKey
+            DbCryptoApiError::InvalidKey
         ));
     }
 
@@ -411,14 +392,14 @@ mod tests {
         struct MyKeyManager {}
         impl KeyManager for MyKeyManager {
             fn get_key(&self) -> ApiResult<Vec<u8>> {
-                Err(LoginsApiError::MissingKey)
+                Err(DbCryptoApiError::MissingKey)
             }
         }
         let key_manager = Arc::new(MyKeyManager {});
         let encdec = ManagedEncryptorDecryptor { key_manager };
         assert!(matches!(
             encdec.encrypt("secret".as_bytes().into()).err().unwrap(),
-            LoginsApiError::MissingKey
+            DbCryptoApiError::MissingKey
         ));
     }
 
@@ -467,7 +448,7 @@ mod tests {
         let bad_key = "bad_key".to_owned();
         assert!(matches!(
             check_canary(&canary, CANARY_TEXT, &bad_key).err().unwrap(),
-            LoginsApiError::InvalidKey
+            DbCryptoApiError::InvalidKey
         ));
     }
 }
@@ -498,7 +479,7 @@ mod tests_keydb {
 
     fn profile_path() -> PathBuf {
         std::path::PathBuf::from(env!("CARGO_MANIFEST_DIR"))
-            .join("../support/rc_crypto/nss/fixtures/profile")
+            .join("../rc_crypto/nss/fixtures/profile")
     }
 
     #[test]
@@ -520,7 +501,10 @@ mod tests_keydb {
         let mock_primary_password_authenticator = MockPrimaryPasswordAuthenticator {
             password: "password".to_string(),
         };
-        let nss_key_manager = NSSKeyManager::new(Arc::new(mock_primary_password_authenticator));
+        let nss_key_manager = NSSKeyManager::new(
+            String::from("as-logins-key"),
+            Arc::new(mock_primary_password_authenticator),
+        );
         // key from fixtures/profile/key4.db
         let expected = [
             123, 34, 107, 116, 121, 34, 58, 34, 111, 99, 116, 34, 44, 34, 107, 34, 58, 34, 66, 74,
@@ -536,9 +520,13 @@ mod tests_keydb {
     fn test_nss_key_manager_caching() {
         ensure_initialized_with_profile_dir(profile_path());
         // `password` is the primary password of the profile fixture
-        let nss_key_manager = NSSKeyManager::new(Arc::new(MockPrimaryPasswordAuthenticator {
+        let mock_primary_password_authenticator = MockPrimaryPasswordAuthenticator {
             password: "password".to_string(),
-        }));
+        };
+        let nss_key_manager = NSSKeyManager::new(
+            String::from("as-logins-key"),
+            Arc::new(mock_primary_password_authenticator),
+        );
 
         let key = nss_key_manager.get_key().unwrap();
         assert_eq!(*nss_key_manager.cached_key.read(), Some(key.clone()));