+++ megazords/fenix-dylib/megazord_stub.c
@@ -13,6 +13,7 @@
 extern int MOZ_EXPORT ffi_ads_client_uniffi_contract_version();
 extern int MOZ_EXPORT ffi_autofill_uniffi_contract_version();
 extern int MOZ_EXPORT ffi_crashtest_uniffi_contract_version();
+extern int MOZ_EXPORT ffi_db_crypto_uniffi_contract_version();
 extern int MOZ_EXPORT ffi_fxa_client_uniffi_contract_version();
 extern int MOZ_EXPORT ffi_init_rust_components_uniffi_contract_version();
 extern int MOZ_EXPORT ffi_logins_uniffi_contract_version();
@@ -77,6 +78,7 @@ void _local_megazord_dummy_symbol() {
     ffi_ads_client_uniffi_contract_version();
     ffi_autofill_uniffi_contract_version();
     ffi_crashtest_uniffi_contract_version();
+    ffi_db_crypto_uniffi_contract_version();
     ffi_fxa_client_uniffi_contract_version();
     ffi_init_rust_components_uniffi_contract_version();
     ffi_logins_uniffi_contract_version();