+++ megazords/full/src/lib.rs
@@ -12,6 +12,7 @@ use std::os::raw::c_char;
 pub use ads_client;
 pub use autofill;
 pub use crashtest;
+pub use db_crypto;
 pub use error_support;
 pub use fxa_client;
 pub use init_rust_components;